# optimize StaticWOperator by using lu to allow saving the factorization