This should fix the regressions caused by https://github.com/SciML/OrdinaryDiffEq.jl/pull/2295 (although I haven't yet fully verified this to my satisfaction).