@oscardssmith 2024-08-14
Benchmarks:
```
using OrdinaryDiffEq, ParameterizedFunctions, LinearAlgebra, StaticArrays

f = @ode_def Orego begin
  dy1 = p1*(y2+y1*(1-p2*y1-y2))
  dy2 = (y3-(1+y1)*y2)/p1
  dy3 = p3*(y1-y3)
end p1 p2 p3

p = SA[77.27,8.375e-6,0.161]
oregostatic = ODEProblem{false}(f,SA[1.0,2.0,3.0],(0.0,30.0),p)


g = @ode_def Hires begin
  dy1 = -1.71*y1 + 0.43*y2 + 8.32*y3 + 0.0007
  dy2 = 1.71*y1 - 8.75*y2
  dy3 = -10.03*y3 + 0.43*y4 + 0.035*y5
  dy4 = 8.32*y2 + 1.71*y3 - 1.12*y4
  dy5 = -1.745*y5 + 0.43*y6 + 0.43*y7
  dy6 = -280.0*y6*y8 + 0.69*y4 + 1.71*y5 -
           0.43*y6 + 0.69*y7
  dy7 = 280.0*y6*y8 - 1.81*y7
  dy8 = -280.0*y6*y8 + 1.81*y7
end

u0 = zeros(8)
u0[1] = 1
u0[8] = 0.0057
hiresstatic = ODEProblem{false}(g,SVector{8}(u0),(0.0,321.8122))

#new 
julia> @btime solve(oregostatic, Rosenbrock23(), dense=false, save_everystep=false);
  64.812 μs (24 allocations: 6.95 KiB))

julia> @btime solve(oregostatic, Rodas5P(), reltol=1e-6, dense=false, save_everystep=false);
  130.234 μs (24 allocations: 8.70 KiB)

julia> @btime solve(hiresstatic, Rosenbrock23(), dense=false, save_everystep=false);
  449.272 μs (3649 allocations: 327.38 KiB)

julia> @btime solve(hiresstatic, Rodas5P(), reltol=1e-6, dense=false, save_everystep=false);
  320.146 μs (2257 allocations: 208.44 KiB)

# old
julia> @btime solve(oregostatic, Rosenbrock23(), dense=false, save_everystep=false);
  61.067 μs (24 allocations: 6.48 KiB)

julia> @btime solve(oregostatic, Rodas5P(), reltol=1e-6, dense=false, save_everystep=false);
  128.326 μs (24 allocations: 8.22 KiB)

julia> @btime solve(hiresstatic, Rosenbrock23(), dense=false, save_everystep=false);
  463.742 μs (3649 allocations: 324.23 KiB)

julia> @btime solve(hiresstatic, Rodas5P(), reltol=1e-6, dense=false, save_everystep=false);
  384.328 μs (2257 allocations: 205.25 KiB)
```

So this is a very minor regression (5%) for the case of an ODE with <8 variables, but a ~20% speedup for 8 or more variables where we are able to reuse the factorization.